+++ src/main/java/com/tbdev/teaneckminyanim/service/TNMUserDetailsService.java
@@ -53,11 +53,18 @@ public class TNMUserDetailsService implements UserDetailsService {
 //            }
 //        }
 
-        UserDetails userDetails = (UserDetails) new User(user.getUsername(), user.getEncryptedPassword(), grantList);
+        UserDetails userDetails = new User(
+                user.getUsername(),
+                user.getEncryptedPassword(),
+                user.isEnabled(),
+                true,
+                true,
+                true,
+                grantList);
 
         return userDetails;
     }
- private String getClientIP() {
+    private String getClientIP() {
         final String xfHeader = request.getHeader("X-Forwarded-For");
         if (xfHeader != null) {
             return xfHeader.split(",")[0];